+++ src/index.scss
@@ -21,6 +21,7 @@ h1 {
 
 .planets-section {
   display: flex;
+  gap: 20px;
   flex-direction: column;
   align-items: center;
 }
@@ -29,54 +30,4 @@ h1 {
   width: 100%;
   display: flex;
   justify-content: space-between;
-}
-
-.pagination__buttons {
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-wrap: wrap;
-  list-style: none;
-}
-
-.pagination__item + .pagination__item {
-  margin-left: 12px;
-}
-
-.pagination__button {
-  font-weight: bold;
-  width: 40px;
-  height: 40px;
-  padding: 10px;
-  border: 0;
-  cursor: pointer;
-  border-radius: 50%;
-  background-color: #CEC7D3;
-
-  &:hover {
-    background-color: #C5B2D4;
-  }
-
-  &:focus {
-    outline: none;
-    box-shadow: 0 0 0 3px #7E569C;
-    background-color: #EAD4FA;
-  }
-
-  &:active {
-    opacity: 0.6;
-    box-shadow: none;
-  }
-}
-
-.pagination__button--active,
-.pagination__button--active:hover,
-.pagination__button--active:focus,
-.pagination__button--active:active
-{
-  color: white;
-  cursor: default;
-  background-color: #795B90;
-  box-shadow: 0 0 0 2px #795B90,
-  inset 0 0 0 2px white;
 }
